+++ src/ReconNess.Web/ClientApp/src/components/target/TargetSubdomains.vue
@@ -19,13 +19,37 @@
 
     <hr />
     <v-client-table :columns="columns" :data="subdomains" :options="options">
-      <a slot="name" slot-scope="props" v-if="props.row.isAlive === true" target="_blank" :href="'http://'+props.row.name" class="glyphicon glyphicon-eye-open">{{props.row.name}}</a>
-      <div slot="name" slot-scope="props" v-else>{{props.row.name}}</div>
-      <div slot="isAlive" slot-scope="props">{{props.row.isAlive}}</div>
-      <div slot="hasHttpOpen" slot-scope="props">{{props.row.hasHttpOpen}}</div>
-      <div slot="actions" slot-scope="props">
-        <button class="btn btn-primary ml-2" v-on:click="onOpenSubdomain(props.row)">Open</button>
-        <button class="btn btn-danger ml-2" v-on:click="onDeleteSubdomain(props.row)">Delete</button>
+      <div slot="name" slot-scope="props" v-if="props.row.isAlive">
+        <a target="_blank" :href="'http://'+props.row.name" class="glyphicon glyphicon-eye-open">{{props.row.name}}</a>        
+      </div>
+      <div slot="name" slot-scope="props" v-else>
+        {{props.row.name}}
+      </div>
+
+      <div class="subdomain-details" slot="details" slot-scope="props">
+        <font-awesome-icon v-if="props.row.isMainPortal" :icon="['fas', 'home']" fixed-width title="Main Portal" />
+        <font-awesome-icon v-if="props.row.isAlive" :icon="['fas', 'heart']" fixed-width title="Alive" />
+        <font-awesome-icon v-if="props.row.hasHttpOpen" :icon="['fas', 'book-open']" fixed-width title="HTTP Open" />
+        <font-awesome-icon v-if="props.row.takeover" :icon="['fas', 'fire-alt']" fixed-width title="Takeover" />
+
+        <div v-if="props.row.fromAgents">Agents: <strong>{{ props.row.fromAgents }} </strong></div>
+        <div v-if="props.row.labels.length > 0" >Labels: <strong v-for="l in props.row.labels" v-bind:key="l.name"><span :style="{ color: l.color}">{{ l.name }} </span></strong></div>
+        <div v-if="props.row.services.length > 0">Services: <strong>{{props.row.services | joinComma('name') }} </strong></div>
+        <div v-if="props.row.ipAddress">IpAddress: <strong>{{props.row.ipAddress }} </strong></div>
+        <div>Added: {{props.row.createdAt | formatDate('YYYY-MM-DD')}}</div>
+      </div>
+
+
+      <div class="subdomain-labels" slot="labels" slot-scope="props">
+        <button type="button" class="btn btn-link" v-on:click="onAddLabel(props.row, 'Checking')" title="Add Checking Label"> <font-awesome-icon :icon="['fas', 'coffee']" /></button>
+        <button type="button" class="btn btn-link" v-on:click="onAddLabel(props.row, 'Vulnerable')" title="Add Vulnerable Label"> <font-awesome-icon :icon="['fas', 'bug']" /></button>
+        <button type="button" class="btn btn-link" v-on:click="onAddLabel(props.row, 'Interesting')" title="Add Interesting Label"> <font-awesome-icon :icon="['fas', 'exclamation']" /></button>
+        <button type="button" class="btn btn-link" v-on:click="onAddLabel(props.row, 'Bounty')" title="Add Bounty Label"> <font-awesome-icon :icon="['fas', 'dollar-sign']" /></button>
+        <button type="button" class="btn btn-link" v-on:click="onAddLabel(props.row, 'Ignore')" title="Add Ignore Label"> <font-awesome-icon :icon="['fas', 'guitar']" /></button>
+      </div>
+      <div class="subdomain-actions" slot="actions" slot-scope="props">
+        <button type="button" class="btn btn-link" v-on:click="onOpenSubdomain(props.row)" title="Open"><font-awesome-icon :icon="['fas', 'arrow-alt-circle-right']" fixed-width /></button>
+        <button type="button" class="btn btn-link" v-on:click="onDeleteSubdomain(props.row)" title="Delete"> <font-awesome-icon :icon="['fas', 'trash-alt']" fixed-width /></button>
       </div>
     </v-client-table>   
   </div>
@@ -45,17 +69,16 @@
       return {
         subdomains: [],
         newSubdomain: null,
-        columns: ['name', 'ipAddress', 'isAlive', 'hasHttpOpen', 'actions'],
+        columns: ['name', 'details', 'labels', 'actions'],
         options: {
           headings: {
             name: 'Subdomain',
-            ipAddress: 'Ip Address',
-            isAlive: 'Is Alive',
-            hasHttpOpen: 'Has Http Open',
-            actions: 'Actions'
+            details: 'Details',
+            labels: 'Labels',
+            actions: 'Actions'            
           },
-          sortable: ['name', 'isAlive'],
-          filterable: ['name', 'isAlive']
+          sortable: ['name'],
+          filterable: ['name']
         }
       }
     },
@@ -77,6 +100,10 @@
           await this.$api.delete('targets/subdomain', this.$route.params.targetName)
           this.subdomains = []
         }
+      },
+      async onAddLabel(subdomain, label) {
+        await this.$api.update('subdomains/label', subdomain.id, { label: label })
+        this.$router.go()
       },
       async onAddNewSubdomain() {
         const target = this.$route.params.targetName
@@ -101,5 +128,13 @@
 
 <!-- Add "scoped" attribute to limit CSS to this component only -->
 <style scoped>
-
+  .subdomain-details {
+    width: 250px;
+}
+  .subdomain-actions {
+    width: 50px;
+}
+  .subdomain-labels {
+    width: 120px;
+}
 </style>