+++ src/ReconNess.Web/Dtos/SubdomainDto.cs
@@ -9,6 +9,8 @@ namespace ReconNess.Web.Dtos
 
         public string Name { get; set; }
 
+        public DateTime CreatedAt { get; set; }
+
         public string Target { get; set; }
 
         public string IpAddress { get; set; }
@@ -27,6 +29,6 @@ namespace ReconNess.Web.Dtos
 
         public List<ServiceDto> Services { get; set; }
 
-        public List<string> Labels { get; set; }
+        public List<LabelDto> Labels { get; set; }
     }
 }