+++ src/ReconNess.Web/Mappers/SubdomainProfile.cs
@@ -15,8 +15,8 @@ namespace ReconNess.Web.Mappers
                 .ForMember(dest => dest.Target, opt => opt.Ignore());
 
             CreateMap<Subdomain, SubdomainDto>()
-                .ForMember(dest => dest.Labels, opt => opt.MapFrom(src => src.Labels.Select(c => c.Label.Name)))
-                .ForMember(dest => dest.Target, opt => opt.MapFrom(src => src.Target.Name));
+                .ForMember(dest => dest.Target, opt => opt.MapFrom(src => src.Target.Name))
+                .ForMember(dest => dest.Labels, opt => opt.MapFrom(src => src.Labels.Select(l => l.Label))); ;
         }
     }
 }